# backend/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoYun API",
        "description": "墨韵在线读书交流平台API文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "开发服务器"
        }
    ],
    "tags": [
        {
            "name": "认证",
            "description": "用户认证相关接口"
        },
        {
            "name": "用户",
            "description": "用户信息相关接口"
        },
        {
            "name": "书评",
            "description": "书评相关接口"
        },
        {
            "name": "圈子",
            "description": "圈子相关接口"
        },
        {
            "name": "消息",
            "description": "消息相关接口"
        },
        {
            "name": "AI助手",
            "description": "AI辅助创作接口"
        },
        {
            "name": "首页",
            "description": "首页相关接口"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": ["认证"],
                "summary": "用户登录",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "account": {
                                        "type": "string",
                                        "description": "用户账号"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "用户密码"
                                    }
                                },
                                "required": ["account", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "account": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": ["认证"],
                "summary": "用户注册",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "account": {
                                        "type": "string",
                                        "description": "用户账号"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "用户密码"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "电子邮箱"
                                    }
                                },
                                "required": ["account", "password", "email"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "注册成功"
                    }
                }
            }
        },
        "/api/users/me": {
            "get": {
                "tags": ["用户"],
                "summary": "获取当前用户信息",
                "responses": {
                    "200": {
                        "description": "成功获取用户信息",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "user": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "account": {
                                                            "type": "string"
                                                        },
                                                        "signature": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["用户"],
                "summary": "更新用户信息",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "signature": {
                                        "type": "string",
                                        "description": "个性签名"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "电子邮箱"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功"
                    }
                }
            }
        },
        "/api/journals": {
            "get": {
                "tags": ["书评"],
                "summary": "获取书评列表",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "每页数量",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "偏移量",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取书评列表"
                    }
                }
            },
            "post": {
                "tags": ["书评"],
                "summary": "发表书评",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "description": "书评标题"
                                    },
                                    "content": {
                                        "type": "string",
                                        "description": "书评内容"
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "description": "书籍ID"
                                    }
                                },
                                "required": ["title", "content", "book_id"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "发表成功"
                    }
                }
            }
        },
        "/api/journals/{id}/like": {
            "post": {
                "tags": ["书评"],
                "summary": "点赞书评",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "书评ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞成功"
                    }
                }
            }
        },
        "/api/groups": {
            "get": {
                "tags": ["圈子"],
                "summary": "获取圈子列表",
                "responses": {
                    "200": {
                        "description": "成功获取圈子列表"
                    }
                }
            },
            "post": {
                "tags": ["圈子"],
                "summary": "创建圈子",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "圈子名称"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "圈子描述"
                                    }
                                },
                                "required": ["name"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "创建成功"
                    }
                }
            }
        },
        "/api/groups/{id}/join": {
            "post": {
                "tags": ["圈子"],
                "summary": "加入圈子",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "圈子ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "加入成功"
                    }
                }
            }
        },
        "/api/messages/chat": {
            "get": {
                "tags": ["消息"],
                "summary": "获取私信列表",
                "responses": {
                    "200": {
                        "description": "成功获取私信列表"
                    }
                }
            },
            "post": {
                "tags": ["消息"],
                "summary": "发送私信",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer",
                                        "description": "接收者ID"
                                    },
                                    "content": {
                                        "type": "string",
                                        "description": "消息内容"
                                    }
                                },
                                "required": ["receiver_id", "content"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "发送成功"
                    }
                }
            }
        },
        "/api/llm/chat": {
            "post": {
                "tags": ["AI助手"],
                "summary": "获取AI回复",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "description": "用户输入"
                                    }
                                },
                                "required": ["message"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功获取回复"
                    }
                }
            }
        },
        "/api/home/dashboard": {
            "get": {
                "tags": ["首页"],
                "summary": "获取首页数据",
                "responses": {
                    "200": {
                        "description": "成功获取首页数据",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "journals": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                },
                                                "weather": {
                                                    "type": "object",
                                                    "properties": {
                                                        "temperature": {
                                                            "type": "number"
                                                        },
                                                        "description": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "unread_message_count": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "SessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        }
    }
}